+++ web/client/js/fetch.js
@@ -14,7 +14,7 @@ async function getUser(token) {
     const {type, access} = token
     if(!access || !type) return false
 
-    return fetch('https://discord.com/api/users/@me', {
+    return fetch('https://discord.com/api/v9/users/@me', {
         headers: {
             authorization: `${type} ${access}`,
             "Content-Type": "application/json"
@@ -29,7 +29,7 @@ async function getGuilds(token) {
     const {type, access} = token
     if(!access || !type) return false
 
-    return fetch('https://discord.com/api/users/@me/guilds', {
+    return fetch('https://discord.com/api/v9/users/@me/guilds', {
         headers: {
             authorization: `${type} ${access}`,
             "Content-Type": "application/json"
@@ -52,14 +52,22 @@ async function getGuilds(token) {
         return filtered
     })
 }
+async function getGuildSettings(guildId) {
+    return await fetch('/getSettings?guild=' + guildId)
+    .then(res => res.json())
+}
+async function getGuildDetails(guildId) {
+    return await fetch('/getGuildDetails?guild=' + guildId)
+    .then(res => res.json())
+    .then(res => {
+        if(res.code === 200) {
+            return res.value
+        }else return false
+    })
+}
 
 async function getTemplate(template) {
     if(!template) return false
 
     return fetch('./templates/' + template).then(res => res.text())
-}
-
-async function getGuildSettings(guildId) {
-    return await fetch('/getSettings?guild=' + guildId)
-    .then(res => res.json())
 }